Major refactoring of Address class and subclasses
The Linear property was eliminated, and each subclass now implements a ToLinear() method that performs appropriate conversion to linear addresses.

Added MakeAddressFromConstant and MakeAddressFromLinear to the Platform class. Anytime an address is created, it should most likely be done through these two methods.

Added a new project, Ps3.csproj, for Playstation3 binaries.

+++ trunk/src/Arch/Cil/CilDisassembler.cs
@@ -889,7 +889,7 @@ namespace Decompiler.Arch.Cil
                 break;
             #endregion
             }
-            instr.Length = rdr.Address - addr;
+            instr.Length = (int)(rdr.Address - addr);
             return instr;
         }
     }
